+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased]
 
+- 🎉 feat: added details button on each neuron + bug fix
 - 🎉 feat: basic layer + neuron blocks added
 <!-- ## [0.0.2] - 2022-12-07